move claimed activity logic into activity module
