+++ packages/shared/src/lib/core/activities/stores/index.ts
@@ -1,2 +1,3 @@
+export * from './claimed-activities.store'
 export * from './persisted-balance-changes.store'
 export * from './persisted-transactions.store'
