+++ packages/shared/src/lib/core/activities/types/claimed-activities.interface.ts
