+++ packages/shared/src/lib/core/activities/types/index.ts
@@ -1,2 +1,3 @@
 export * from './asset-balance-change.interface'
+export * from './claimed-activities.interface'
 export * from './persisted-evm-transaction.interface'
