+++ packages/shared/src/lib/core/wallet/interfaces/index.ts
@@ -7,7 +7,6 @@ export * from './asset-balance.interface'
 export * from './asset-filter.interface'
 export * from './asset.interface'
 export * from './base-token.interface'
-export * from './claimed-activities.interface'
 export * from './contact-subject.interface'
 export * from './erc20-metadata.interface'
 export * from './hidden-activities.interface'
